fix hpo tests
