+++ multimodal/tests/unittests/test_predictor.py
@@ -5,12 +5,9 @@ import numpy.testing as npt
 import tempfile
 import copy
 import pickle
-
 from omegaconf import OmegaConf
-from ray import tune
 from torch import nn
 
-from autogluon.core.hpo.constants import SEARCHER_PRESETS, SCHEDULER_PRESETS
 from autogluon.multimodal import AutoMMPredictor
 from autogluon.multimodal.utils import modify_duplicate_model_names
 from autogluon.multimodal.constants import (
@@ -588,6 +585,7 @@ def test_modifying_duplicate_model_names():
         for per_processor in per_modality_processors:
             assert per_processor.prefix in teacher_predictor._config.model.names
 
+
 def test_mixup():
     dataset = ALL_DATASETS["petfinder"]()
     metric_name = dataset.metric
@@ -624,6 +622,7 @@ def test_mixup():
 
         score = predictor.evaluate(dataset.test_df)
         verify_predictor_save_load(predictor, dataset.test_df)
+
 
 def test_textagumentor_deepcopy():
     dataset = ALL_DATASETS["ae"]()
@@ -685,135 +684,6 @@ def test_textagumentor_deepcopy():
         time_limit=10,
     )
 
-@pytest.mark.parametrize('searcher', list(SEARCHER_PRESETS.keys()))
-@pytest.mark.parametrize('scheduler', list(SCHEDULER_PRESETS.keys()))
-def test_hpo(searcher, scheduler):
-    dataset = PetFinderDataset()
-
-    config = {
-        MODEL: f"fusion_mlp_image_text_tabular",
-        DATA: "default",
-        OPTIMIZATION: "adamw",
-        ENVIRONMENT: "default",
-    }
-
-    hyperparameters = {
-        "optimization.learning_rate": tune.uniform(0.0001, 0.01),
-        "optimization.max_epochs": 1,
-        "model.names": ["numerical_mlp", "categorical_mlp", "fusion_mlp"],
-        "env.num_workers": 0,
-        "env.num_workers_evaluation": 0,
-    }
-
-    hyperparameter_tune_kwargs = {
-        'searcher': searcher,
-        'scheduler': scheduler,
-        'num_trials': 2,
-    }
-
-    predictor = AutoMMPredictor(
-        label=dataset.label_columns[0],
-        problem_type=dataset.problem_type,
-        eval_metric=dataset.metric,
-    )
-
-    save_path = os.path.join(get_home_dir(), 'hpo', f'_{searcher}', f'_{scheduler}')
-    if os.path.exists(save_path):
-        shutil.rmtree(save_path)
-
-    predictor = predictor.fit(
-        train_data=dataset.train_df,
-        config=config,
-        hyperparameters=hyperparameters,
-        time_limit=60,
-        save_path=save_path,
-        hyperparameter_tune_kwargs=hyperparameter_tune_kwargs,
-    )
-
-    score = predictor.evaluate(dataset.test_df)
-    verify_predictor_save_load(predictor, dataset.test_df)
-
-    # test for continuous training
-    predictor = predictor.fit(
-        train_data=dataset.train_df,
-        config=config,
-        hyperparameters=hyperparameters,
-        time_limit=60,
-        hyperparameter_tune_kwargs=hyperparameter_tune_kwargs,
-    )
-
-
-@pytest.mark.parametrize('searcher', list(SEARCHER_PRESETS.keys()))
-@pytest.mark.parametrize('scheduler', list(SCHEDULER_PRESETS.keys()))
-def test_hpo_distillation(searcher, scheduler):
-    dataset = PetFinderDataset()
-
-    config = {
-        MODEL: f"fusion_mlp_image_text_tabular",
-        DATA: "default",
-        OPTIMIZATION: "adamw",
-        ENVIRONMENT: "default",
-        DISTILLER: "default",
-    }
-
-    hyperparameters = {
-        "optimization.max_epochs": 1,
-        "model.names": ["numerical_mlp", "categorical_mlp", "fusion_mlp"],
-        "env.num_workers": 0,
-        "env.num_workers_evaluation": 0,
-    }
-
-    hyperparameter_tune_kwargs = {
-        'searcher': searcher,
-        'scheduler': scheduler,
-        'num_trials': 2,
-    }
-
-    teacher_predictor = AutoMMPredictor(
-        label=dataset.label_columns[0],
-        problem_type=dataset.problem_type,
-        eval_metric=dataset.metric,
-    )
-
-    teacher_save_path = os.path.join(get_home_dir(), 'hpo_distillation_teacher', f'_{searcher}', f'_{scheduler}')
-    if os.path.exists(teacher_save_path):
-        shutil.rmtree(teacher_save_path)
-
-    teacher_predictor = teacher_predictor.fit(
-        train_data=dataset.train_df,
-        config=config,
-        hyperparameters=hyperparameters,
-        time_limit=60,
-        save_path=teacher_save_path,
-    )
-
-    hyperparameters = {
-        "optimization.learning_rate": tune.uniform(0.0001, 0.01),
-        "optimization.max_epochs": 1,
-        "model.names": ["numerical_mlp"],
-        "env.num_workers": 0,
-        "env.num_workers_evaluation": 0,
-    }
-    # test for distillation
-    predictor = AutoMMPredictor(
-        label=dataset.label_columns[0],
-        problem_type=dataset.problem_type,
-        eval_metric=dataset.metric,
-    )
-
-    student_save_path = os.path.join(get_home_dir(), 'hpo_distillation_student', f'_{searcher}', f'_{scheduler}')
-    if os.path.exists(student_save_path):
-        shutil.rmtree(student_save_path)
-
-    predictor = predictor.fit(
-        train_data=dataset.train_df,
-        teacher_predictor=teacher_save_path,
-        config=config,
-        hyperparameters=hyperparameters,
-        time_limit=60,
-        hyperparameter_tune_kwargs=hyperparameter_tune_kwargs,
-        save_path=student_save_path,
-    )
 
 def test_trivialaugment():
     dataset = ALL_DATASETS["petfinder"]()
